grafana: README: Add information for split access to grafana
Note: Grafana is a web server of it's own. Now, we route:
a) https://site/path which is basically authenticated with
usual apache authentication scheme, which is also used to login into
grafana -> this can be exported as a port over your router

b) http://internalnetwork/path/ which does not have apache
authentication, useful for screensavers and other internal webpage
usage.

Signed-off-by: Nishanth Menon <menon.nishanth@gmail.com>